requirements.txt
requests
bs4
python-dotenv
groq
neo4j
streamlit
plotly
langchain
langchain_groq
tiktoken
streamlit
matplotlib
networkx
flask